% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{Call Image Colorization API}
\usage{
colorize(
  img,
  key = "quickstart-QUdJIGlzIGNvbWluZy4uLi4K",
  pane = c("plot", "view", "none")
)
}
\arguments{
\item{img}{path or URL to the images to be colorized.}

\item{key}{DeepAI API key.}

\item{pane}{defines in which pane the image is displayed: \code{plot}, \code{view} or \code{none}.}
}
\value{
A tibble with the file locations of the original images and the response URLs to
   the colorized images. The latter expire after some time. It is recommended to save the
   colorized images.
}
\description{
\code{colorize} calls the DeOldify image colorization API on DeepAI.
}
\details{
With the default API key a few queries are possible. After registration on DeepAI
   \url{https://deepai.org/}, around 5000 free requests are currently possible.
}
\examples{
\dontrun{
# Call image colorization API
colorize(img = "https://upload.wikimedia.org/wikipedia/commons/9/9e/Breadfruit.jpg")
}
}
