% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorizer.R
\docType{package}
\name{colorizer}
\alias{colorizer}
\title{\code{colorizer} package}
\description{
Colorize and Restore Old Images Using the 'DeOldify' Image Colorization API on 'DeepAI'
}
\details{
See the README on
\href{https://github.com/zumbov2/colorizer#readme}{GitHub}
}
